refactor: relocate Kalman filter module
